refactored index.js and put place bet supporting functions in utility-functions folder
