+++ client/pages/account-page.jsx
@@ -4,16 +4,12 @@ import Oddsbar from '../components/odds-bar';
 import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
-import Table from 'react-bootstrap/Table';
 import PlaceholderTable from '../components/placeholder';
 import AppContext from '../lib/app-context';
-import { abbreviationsObject } from '../lib/abbreviations';
-import Button from 'react-bootstrap/Button';
-import Modal from 'react-bootstrap/Modal';
-import Form from 'react-bootstrap/Form';
-import InputGroup from 'react-bootstrap/InputGroup';
 import Footer from '../components/footer';
 import Redirect from '../components/redirect';
+import BetHistoryTable from '../components/bet-history-table';
+import DepositModal from '../components/deposit-modal';
 
 export default class AccountPage extends React.Component {
   constructor(props) {
@@ -103,7 +99,7 @@ export default class AccountPage extends React.Component {
     if (depositAmount > 10000) {
       return { errorMessage: 'Deposit amount must be less than $10,000' };
     }
-    if (depositAmount + this.context.accountBalance > 10000) {
+    if (depositAmount + this.state.accountBalance > 10000) {
       return { errorMessage: 'Deposit and current account balance not to exceed $10,000' };
     }
     return {};
@@ -115,6 +111,7 @@ export default class AccountPage extends React.Component {
     if (Object.keys(errorsObject).length > 0) {
       this.setState({
         errorMessage: errorsObject.errorMessage,
+        depositAmount: this.state.depositAmount,
         show: true
       });
     } else {
@@ -197,96 +194,20 @@ export default class AccountPage extends React.Component {
                 <a onClick={this.handleShow} id="deposit-anchor">Running Low on Funds?</a>
               </Col>
             </Row>
-            <Table bordered className='mt-5' id='bet-history-table' fluid="md">
-              <thead>
-                <tr className="td-no-wrap">
-                  <th className='align-middle table-data-20'>Placed Date</th>
-                  <th className="table-data-40">Bet</th>
-                  <th className="table-data-20">Amount</th>
-                  <th className="table-data-20">Status</th>
-                </tr>
-              </thead>
-              <tbody>
-                {
-                  this.state.betHistory.map(elem => {
-                    let homeTeamScoreColor;
-                    let awayTeamScoreColor;
-                    if (elem.homeTeamScore > elem.awayTeamScore) {
-                      homeTeamScoreColor = 'winning-score';
-                      awayTeamScoreColor = 'losing-score';
-                    } else {
-                      awayTeamScoreColor = 'winning-score';
-                      homeTeamScoreColor = 'losing-score';
-                    }
-                    let betStatusColor;
-                    let operator;
-                    if (elem.status === 'won') {
-                      betStatusColor = 'green-color';
-                      operator = '+';
-                    } else if (elem.status === 'lost') {
-                      betStatusColor = 'red-color';
-                      operator = '-';
-                    } else {
-                      betStatusColor = 'white-color';
-                      operator = '-';
-                    }
-                    return (
-                      <tr className='td-no-wrap td-quarter' key={elem.betId}>
-                        <td className="align-middle">{new Date(elem.createdAt).toLocaleDateString()}</td>
-                        <td className="double-line-height">
-                          <span id="bet-history-game-details">
-                            {new Date(elem.gameStart).toLocaleDateString()}: {new Date(elem.gameStart).toLocaleTimeString()}
-                            <br />
-                            <span className='abbreviated-text'>{abbreviationsObject[elem.awayTeam]}</span>
-                            <span className='full-text'>{elem.awayTeam}</span>
-                            : <span className={awayTeamScoreColor}>{elem.awayTeamScore} </span>
-                            @
-                            <span className='abbreviated-text'> {abbreviationsObject[elem.homeTeam]}</span>
-                            <span className='full-text'> {elem.homeTeam}</span>
-                            : <span className={homeTeamScoreColor}>{elem.homeTeamScore}</span>
-                            <br />
-                          </span>
-                          <span className='abbreviated-text'> {abbreviationsObject[elem.winningTeam]} </span>
-                          <span className='full-text'> {elem.winningTeam} </span>
-                          {elem.betType.charAt(0).toUpperCase() + elem.betType.slice(1)} {elem.points > 0 ? '+' : ''}{elem.points} ({elem.price > 0 ? '+' : ''}{elem.price})
-                        </td>
-                        <td>
-                          <span className={betStatusColor}>{operator}{new Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD' }).format(betStatusColor === 'green-color' ? elem.betAmount + elem.potentialWinnings : elem.betAmount)}</span>
-                        </td>
-                        <td>
-                          <span className={betStatusColor}>{elem.status}</span>
-                        </td>
-                      </tr>
-                    );
-                  })
-                }
-              </tbody>
-            </Table>
+            <BetHistoryTable betHistory={this.state.betHistory} />
           </Container>
         </div>
         <Footer className="footer" />
-        <Modal show={this.state.show} onHide={this.handleClose}>
-          <Modal.Header closeButton>
-            <Modal.Title>Deposit More $$$</Modal.Title>
-          </Modal.Header>
-          <Modal.Body>
-            <Form noValidate validated={this.state.valid} onSubmit={this.handleSubmit}>
-              <p><strong>Deposit anywhere between $1 and $10,000</strong></p>
-              <p>Account balance not to exceed $10,000</p>
-              <Form.Group className="mb-3">
-                <Form.Label>Deposit Amount:</Form.Label>
-                <InputGroup>
-                  <InputGroup.Text>$</InputGroup.Text>
-                  <Form.Control required type="text" onChange={this.handleChange} value={this.state.depositAmount} isInvalid={!!this.state.errorMessage} />
-                  <Form.Control.Feedback type="invalid">{this.state.errorMessage}</Form.Control.Feedback>
-                </InputGroup>
-              </Form.Group>
-              <Button type="submit" id="deposit-submit-button">
-                $$$
-              </Button>
-            </Form>
-          </Modal.Body>
-        </Modal>
+        <DepositModal
+          show={this.state.show}
+          onHide={this.handleClose}
+          validated={this.state.valid}
+          handleSubmit={this.handleSubmit}
+          onChange={this.handleChange}
+          value={this.state.depositAmount}
+          isInvalid={!!this.state.errorMessage}
+          errorMessage={this.state.errorMessage}
+        />
       </>
     );
   }
