+++ server/utility-functions/check-deposit.js
@@ -1,4 +1,4 @@
-function checkDeposit(depositAmount) {
+function checkDeposit(depositAmount, accountBalance) {
   if (isNaN(depositAmount)) {
     return 'Deposit amount must be a valid number';
   }
@@ -8,6 +8,9 @@ function checkDeposit(depositAmount) {
   }
   if (depositAmount > 10000) {
     return 'Deposit amount must be less than $10,000';
+  }
+  if (depositAmount + accountBalance > 10000) {
+    return 'Account balance cannot exceed $10,000';
   }
   return true;
 }
